reader_consumer/pyproject.toml
[tool.poetry]
name = "pipelinenodes"
version = "0.1.0"
description = ""
authors = ["aleksej <xfrbt1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.6.1"
pydantic-settings = "^2.1.0"

kafka-python = "^2.0.2"

pymongo = "^4.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"